+++ src/main/java/sql/service/ServiceMySQL.java
@@ -2,12 +2,12 @@ package sql.service;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import sql.dao.mysql.CityDaoMySQL;
+import sql.dao.mysql.CityDAO;
 import sql.dao.ICityDAO;
-import sql.dao.mysql.CountryDaoMySQL;
+import sql.dao.mysql.CountryDAO;
 import sql.dao.ICountryDAO;
 import sql.dao.IManufacturerDAO;
-import sql.dao.mysql.ManufacturerDaoMySQL;
+import sql.dao.mysql.ManufacturerDAO;
 import sql.datamodels.entity.City;
 import sql.datamodels.entity.Country;
 import sql.datamodels.entity.Manufacturer;
@@ -15,9 +15,9 @@ import sql.datamodels.entity.Manufacturer;
 import java.util.List;
 
 public class ServiceMySQL implements IService { // String url, String user, String pass, int size
-    private static final ICityDAO cityDao = new CityDaoMySQL();
-    private static final ICountryDAO countryDao = new CountryDaoMySQL();
-    private static final IManufacturerDAO manufacturerDao = new ManufacturerDaoMySQL();
+    private static final ICityDAO cityDao = new CityDAO();
+    private static final ICountryDAO countryDao = new CountryDAO();
+    private static final IManufacturerDAO manufacturerDao = new ManufacturerDAO();
     private static final Logger LOGGER = LogManager.getLogger("TESTLOGGER");
 
     @Override
